Updates
* Remove commented code
* Move process switch stuff to a correct header
* Reorder process switch arguments
